Inlined unused GLSLDeclaratorBase
