+++ src/glslplugin/lang/elements/declarations/GLSLDeclarator.java
@@ -20,12 +20,16 @@
 package glslplugin.lang.elements.declarations;
 
 import com.intellij.lang.ASTNode;
+import com.intellij.psi.PsiElement;
+import glslplugin.lang.elements.GLSLElementImpl;
 import glslplugin.lang.elements.GLSLIdentifier;
 import glslplugin.lang.elements.expressions.GLSLExpression;
+import glslplugin.lang.elements.types.GLSLArrayType;
+import glslplugin.lang.elements.types.GLSLQualifiedType;
+import glslplugin.lang.elements.types.GLSLType;
+import glslplugin.lang.elements.types.GLSLTypes;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-
-import java.util.logging.Logger;
 
 /**
  * GLSLDeclarator represents a local or global variable declaration.
@@ -35,7 +39,7 @@ import java.util.logging.Logger;
  *         Date: Jan 29, 2009
  *         Time: 7:29:46 PM
  */
-public class GLSLDeclarator extends GLSLDeclaratorBase {
+public class GLSLDeclarator extends GLSLElementImpl {
     public static final GLSLDeclarator[] NO_DECLARATORS = new GLSLDeclarator[0];
 
     public GLSLDeclarator(@NotNull ASTNode astNode) {
@@ -53,7 +57,75 @@ public class GLSLDeclarator extends GLSLDeclaratorBase {
     }
 
     @Nullable
-    public GLSLArraySpecifier getArraySpecifier() {
-        return findChildByClass(GLSLArraySpecifier.class);
+    public GLSLIdentifier getIdentifier() {
+        PsiElement idElement = getFirstChild();
+        if (idElement instanceof GLSLIdentifier) {
+            return (GLSLIdentifier) idElement;
+        } else {
+            return null; //May trigger on malformed code
+        }
+    }
+
+    @NotNull
+    public String getIdentifierName() {
+        PsiElement idElement = getFirstChild();
+        if (idElement instanceof GLSLIdentifier) {
+            return ((GLSLIdentifier) idElement).getIdentifierName();
+        } else {
+            return "(anonymous)";
+        }
+    }
+
+    @Nullable
+    public GLSLDeclaration getParentDeclaration() {
+        return findParentByClass(GLSLDeclarationImpl.class);
+    }
+
+    @NotNull
+    public GLSLType getType() {
+        GLSLDeclaration declaration = getParentDeclaration();
+        if(declaration == null)return GLSLTypes.UNKNOWN_TYPE;
+        GLSLTypeSpecifier declarationType = declaration.getTypeSpecifierNode();
+        if(declarationType == null)return GLSLTypes.UNKNOWN_TYPE;
+
+        GLSLType declaredType = declarationType.getType();
+        if(!declaredType.isValidType())return GLSLTypes.UNKNOWN_TYPE;
+
+        GLSLArraySpecifier[] arraySpecifiers = findChildrenByClass(GLSLArraySpecifier.class);
+        if(arraySpecifiers.length == 0){
+            return declaredType;
+        }else{
+            //Must append some dimensions to the type
+            if(declaredType instanceof GLSLArrayType){
+                //Already an array, must append the dimensions
+                GLSLArrayType declaredArrayType = (GLSLArrayType) declaredType;
+                int[] existingDimensions = declaredArrayType.getDimensions();
+                int[] combinedDimensions = new int[existingDimensions.length + arraySpecifiers.length];
+                System.arraycopy(existingDimensions, 0, combinedDimensions, 0, existingDimensions.length);
+                for (int i = 0; i < arraySpecifiers.length; i++) {
+                    combinedDimensions[i + existingDimensions.length] = arraySpecifiers[i].getDimensionSize();
+                }
+                return new GLSLArrayType(declaredArrayType.getBaseType(), combinedDimensions);
+            }else{
+                int[] dimensions = new int[arraySpecifiers.length];
+                for (int i = 0; i < dimensions.length; i++) {
+                    dimensions[i] = arraySpecifiers[i].getDimensionSize();
+                }
+                return new GLSLArrayType(declaredType, dimensions);
+            }
+        }
+    }
+
+    @NotNull
+    public GLSLQualifiedType getQualifiedType() {
+        final GLSLType type = getType();
+        final GLSLDeclaration declaration = getParentDeclaration();
+        if(declaration == null || declaration.getQualifierList() == null)return new GLSLQualifiedType(type);
+        return new GLSLQualifiedType(type, declaration.getQualifierList().getQualifiers());
+    }
+
+    @Override
+    public String toString() {
+        return "Declarator: " + getIdentifierName() + " : " + getType().getTypename();
     }
 }
